Simplification, cleanup.
