+++ packages/bamjs/vite.config.js
@@ -1,10 +1,22 @@
-import vue from "@vitejs/plugin-vue";
 import { defineConfig } from "vite";
-
-import { viteConfigCharts } from "./vite.config.charts";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-    plugins: [vue()],
-    ...viteConfigCharts,
+    plugins: [],
+    build: {
+        outDir: "./static/dist",
+        emptyOutDir: true,
+        rollupOptions: {
+            output: {
+                entryFileNames: "[name].js",
+                chunkFileNames: "[name].js",
+                assetFileNames: "[name][extname]",
+            },
+        },
+    },
+    resolve: {
+        alias: {
+            "@": "/src",
+        },
+    },
 });
