+++ packages/bamjs/vitest.config.js
@@ -1,8 +1,7 @@
 import { defineConfig } from "vitest/config";
-import vue from "@vitejs/plugin-vue";
 
 export default defineConfig({
-    plugins: [vue()],
+    plugins: [],
     test: {
         environment: "jsdom",
         globals: true,
